+++ lib/src/models/request/a2a_eft_payment_request.dart
@@ -22,7 +22,6 @@ sealed class PPEftPaymentRequestModel with _$PPEftPaymentRequestModel {
     required PPPaymentMethod paymentMethod,
     required String transactionId,
     required int taxRate,
-    PPDeviceTypeEnum deviceType = PPDeviceTypeEnum.POS,
     int? installment,
   }) => PPEftPaymentRequestModel(
     data: PPEftPaymentRequestData(
@@ -36,7 +35,6 @@ sealed class PPEftPaymentRequestModel with _$PPEftPaymentRequestModel {
     header: PPGeneralRequestHeader(
       transactionType: PPTransactionType.POST_EFTPOS,
       clientToken: clientToken,
-      deviceType: deviceType
     ),
   );
 }